+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/FastInitBoxTestGenerated.java
@@ -1566,12 +1566,6 @@ public class FastInitBoxTestGenerated extends AbstractFastInitBoxTest {
       }
 
       @Test
-      @TestMetadata("AnotherCustomIdentifiersTest.kt")
-      public void testAnotherCustomIdentifiersTest() {
-        runTest("compiler-tests/src/test/data/box/inject/assisted/AnotherCustomIdentifiersTest.kt");
-      }
-
-      @Test
       @TestMetadata("AssistedAnnotationsUseParamNameByDefault.kt")
       public void testAssistedAnnotationsUseParamNameByDefault() {
         runTest("compiler-tests/src/test/data/box/inject/assisted/AssistedAnnotationsUseParamNameByDefault.kt");